docs: Revamp README and enhance project setup instructions
- Updated project name to "RestJam - Restaurant Social Media Application".
- Expanded project description and added tech stack details.
- Improved installation instructions with clear steps for setting up environment variables.
- Added prerequisites for running the application.
- Updated server start script in package.json for better development experience.
- Refactored GraphQL resolvers to utilize service functions for cleaner code.
