+++ server/services/index.js
@@ -1,0 +1,6 @@
+// Services index - Export all services for easy importing
+module.exports = {
+  userService: require('./userService'),
+  postService: require('./postService'),
+  postFieldService: require('./postFieldService'),
+}
